Move some more color conversion functions to ColorConv
