+++ GPU/Common/TextureDecoder.cpp
@@ -328,124 +328,3 @@ void DecodeDXT5Block(u32 *dst, const DXT5Block *src, int pitch) {
 		dst += pitch;
 	}
 }
-
-void ConvertBGRA8888ToRGBA8888(u32 *dst, const u32 *src, const u32 numPixels) {
-#ifdef _M_SSE
-	const __m128i maskGA = _mm_set1_epi32(0xFF00FF00);
-
-	const __m128i *srcp = (const __m128i *)src;
-	__m128i *dstp = (__m128i *)dst;
-	u32 sseChunks = numPixels / 4;
-	if (((intptr_t)src & 0xF) || ((intptr_t)dst & 0xF)) {
-		sseChunks = 0;
-	}
-	for (u32 i = 0; i < sseChunks; ++i) {
-		__m128i c = _mm_load_si128(&srcp[i]);
-		__m128i rb = _mm_andnot_si128(maskGA, c);
-		c = _mm_and_si128(c, maskGA);
-
-		__m128i b = _mm_srli_epi32(rb, 16);
-		__m128i r = _mm_slli_epi32(rb, 16);
-		c = _mm_or_si128(_mm_or_si128(c, r), b);
-		_mm_store_si128(&dstp[i], c);
-	}
-	// The remainder starts right after those done via SSE.
-	u32 i = sseChunks * 4;
-#else
-	u32 i = 0;
-#endif
-	for (; i < numPixels; i++) {
-		const u32 c = src[i];
-		dst[i] = ((c >> 16) & 0x000000FF) |
-		         ((c >> 0)  & 0xFF00FF00) |
-		         ((c << 16) & 0x00FF0000);
-	}
-}
-
-void ConvertRGBA8888ToRGBA5551(u16 *dst, const u32 *src, const u32 numPixels) {
-#if _M_SSE >= 0x401
-	const __m128i maskAG = _mm_set1_epi32(0x8000F800);
-	const __m128i maskRB = _mm_set1_epi32(0x00F800F8);
-	const __m128i mask = _mm_set1_epi32(0x0000FFFF);
-
-	const __m128i *srcp = (const __m128i *)src;
-	__m128i *dstp = (__m128i *)dst;
-	u32 sseChunks = (numPixels / 4) & ~1;
-	// SSE 4.1 required for _mm_packus_epi32.
-	if (((intptr_t)src & 0xF) || ((intptr_t)dst & 0xF) || !cpu_info.bSSE4_1) {
-		sseChunks = 0;
-	}
-	for (u32 i = 0; i < sseChunks; i += 2) {
-		__m128i c1 = _mm_load_si128(&srcp[i + 0]);
-		__m128i c2 = _mm_load_si128(&srcp[i + 1]);
-		__m128i ag, rb;
-
-		ag = _mm_and_si128(c1, maskAG);
-		ag = _mm_or_si128(_mm_srli_epi32(ag, 16), _mm_srli_epi32(ag, 6));
-		rb = _mm_and_si128(c1, maskRB);
-		rb = _mm_or_si128(_mm_srli_epi32(rb, 3), _mm_srli_epi32(rb, 9));
-		c1 = _mm_and_si128(_mm_or_si128(ag, rb), mask);
-
-		ag = _mm_and_si128(c2, maskAG);
-		ag = _mm_or_si128(_mm_srli_epi32(ag, 16), _mm_srli_epi32(ag, 6));
-		rb = _mm_and_si128(c2, maskRB);
-		rb = _mm_or_si128(_mm_srli_epi32(rb, 3), _mm_srli_epi32(rb, 9));
-		c2 = _mm_and_si128(_mm_or_si128(ag, rb), mask);
-
-		_mm_store_si128(&dstp[i / 2], _mm_packus_epi32(c1, c2));
-	}
-	// The remainder starts right after those done via SSE.
-	u32 i = sseChunks * 4;
-#else
-	u32 i = 0;
-#endif
-	for (; i < numPixels; i++) {
-		dst[i] = RGBA8888toRGBA5551(src[i]);
-	}
-}
-
-inline u16 BGRA8888toRGBA5551(u32 px) {
-	return ((px >> 19) & 0x001F) | ((px >> 6) & 0x03E0) | ((px << 7) & 0x7C00) | ((px >> 16) & 0x8000);
-}
-
-void ConvertBGRA8888ToRGBA5551(u16 *dst, const u32 *src, const u32 numPixels) {
-#if _M_SSE >= 0x401
-	const __m128i maskAG = _mm_set1_epi32(0x8000F800);
-	const __m128i maskRB = _mm_set1_epi32(0x00F800F8);
-	const __m128i mask = _mm_set1_epi32(0x0000FFFF);
-
-	const __m128i *srcp = (const __m128i *)src;
-	__m128i *dstp = (__m128i *)dst;
-	u32 sseChunks = (numPixels / 4) & ~1;
-	// SSE 4.1 required for _mm_packus_epi32.
-	if (((intptr_t)src & 0xF) || ((intptr_t)dst & 0xF) || !cpu_info.bSSE4_1) {
-		sseChunks = 0;
-	}
-	for (u32 i = 0; i < sseChunks; i += 2) {
-		__m128i c1 = _mm_load_si128(&srcp[i + 0]);
-		__m128i c2 = _mm_load_si128(&srcp[i + 1]);
-		__m128i ag, rb;
-
-		ag = _mm_and_si128(c1, maskAG);
-		ag = _mm_or_si128(_mm_srli_epi32(ag, 16), _mm_srli_epi32(ag, 6));
-		rb = _mm_and_si128(c1, maskRB);
-		rb = _mm_or_si128(_mm_srli_epi32(rb, 19), _mm_slli_epi32(rb, 7));
-		c1 = _mm_and_si128(_mm_or_si128(ag, rb), mask);
-
-		ag = _mm_and_si128(c2, maskAG);
-		ag = _mm_or_si128(_mm_srli_epi32(ag, 16), _mm_srli_epi32(ag, 6));
-		rb = _mm_and_si128(c2, maskRB);
-		rb = _mm_or_si128(_mm_srli_epi32(rb, 19), _mm_slli_epi32(rb, 7));
-		c2 = _mm_and_si128(_mm_or_si128(ag, rb), mask);
-
-		_mm_store_si128(&dstp[i / 2], _mm_packus_epi32(c1, c2));
-	}
-	// The remainder starts right after those done via SSE.
-	u32 i = sseChunks * 4;
-#else
-	u32 i = 0;
-#endif
-	for (; i < numPixels; i++) {
-		dst[i] = BGRA8888toRGBA5551(src[i]);
-	}
-}
